restructure query in common
