+++ src/core/navbar/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import Navbar from './navbar';
-import { receiveQuery } from 'actions/query_actions'; // TODO: Update!
+import { receiveQuery } from 'common/query/actions';
 import { loginUser, receiveUser } from 'common/modules/user/actions';
 
 const mapStateToProps = ({ user }) => ({
