+++ src/modules/subscriptions/actions.js
@@ -1,4 +1,4 @@
-import * as YoutubeVideoAPI from 'util/youtube_video_util';
+import * as YoutubeVideoAPI from './util';
 
 export const RECEIVE_SUBSCRIPTIONS = 'RECEIVE_SUBSCRIPTIONS';
 export const RECEIVE_SUBSCRIPTIONS_UPLOADS = 'RECEIVE_SUBSCRIPTIONS_UPLOADS';
