+++ src/modules/subscriptions/util.js
@@ -1,4 +1,4 @@
-import * as YoutubeApi from '../youtube_api';
+import * as YoutubeApi from 'core/youtube_api';
 
 export const fetchAuthUserSubscriptions = channelId => {
   let params = {
