--- src3/actions/setting_actions.js
@@ -1,6 +1,0 @@
-// export const RECEIVE_SETTING = 'RECEIVE_SETTING';
-//
-// export const receiveSetting = setting => ({
-//   type: RECEIVE_SETTING,
-//   setting
-// });
