--- src3/reducers/root_reducer.js
@@ -1,28 +1,0 @@
-import { combineReducers } from 'redux';
-import channelsReducer from './channels_reducer';
-// import channelDetailsReducer from './channel_details_reducer';
-import queryReducer from './query_reducer';
-import recommendedReducer from './recommended_reducer';
-import searchResultReducer from './search_result_reducer';
-import subscriptionsReducer from './subscriptions_reducer';
-// import trendingReducer from './trending_reducer';
-// import userReducer from './user_reducer';
-// import settingReducer from './setting_reducer';
-import playlistsReducer from './playlists_reducer';
-import playerDetailsReducer from './player_details_reducer';
-
-const rootReducer = combineReducers({
-  channels: channelsReducer,
-  // channelDetails: channelDetailsReducer,
-  query: queryReducer,
-  recommended: recommendedReducer,
-  searchResult: searchResultReducer,
-  subscriptions: subscriptionsReducer,
-  // trending: trendingReducer,
-  // user: userReducer,
-  // setting: settingReducer,
-  playlists: playlistsReducer,
-  playerDetails: playerDetailsReducer
-});
-
-export default rootReducer;
